+++ src/ckanapi_harvesters/auxiliary/custom/__init__.py
@@ -1,0 +1,8 @@
+#!python3
+# -*- coding: utf-8 -*-
+"""
+Auxiliary files for external use
+"""
+
+from . import aux_helpers
+from . import shared_params_2030