+++ src/ckanapi_harvesters/auxiliary/custom/shared_params_2030.py
@@ -1,0 +1,17 @@
+#!python3
+# -*- coding: utf-8 -*-
+"""
+Definitions used externally (not used in current package)
+"""
+from ckanapi_harvesters.auxiliary.custom.aux_helpers import replace_prefix
+
+
+harvest_group_prefix = "catalogue_"
+private_group_prefix = "private_"
+
+auth_role_ckan_group_prefix = "CKAN_GROUP_"
+
+access_terms_field = "Access Terms"
+access_terms_sep_re = r"[;\n]+"
+
+