+++ src/ckanapi_harvesters/builder/builder_package_1_basic.py
@@ -1491,6 +1491,7 @@ class BuilderPackageBasic:
 
     def setup_sample_package(self, ckan: CkanApi, package_name:str=None, *,
                              sample_url_suffix:str=None, sample_title_suffix:str=None,
+                             sample_groups:List[str]=None,
                              sample_df_dict: Dict[str, Union[bytes, pd.DataFrame]]=None, return_sample:bool=False,
                              records_to_file:DataStoreReprFormat=None, **kwargs) \
         -> Union["BuilderPackageBasic", Tuple["BuilderPackageBasic", Dict[str, Union[bytes, pd.DataFrame]]]]:
@@ -1502,6 +1503,7 @@ class BuilderPackageBasic:
         :param package_name: If specified, derives the package metadata from the specified package name. By default, the current package builder will be used.
         :param sample_url_suffix: Suffix to add to the package_name (default is "-sample")
         :param sample_title_suffix: Suffix to add to the package title (default is " - Sample")
+        :param sample_groups: Option to replace the groups from the original package with a new list of group names
         :param sample_df_dict: Option to transmit the data of each resource to the output of function.
         :param return_sample: Option to return the data of each resource.
         :param kwargs: Optional arguments to pass to the download_sample function.
@@ -1526,6 +1528,8 @@ class BuilderPackageBasic:
         sample_mdl.package_attributes.title = mdl.package_attributes.title + sample_title_suffix
         sample_mdl.package_attributes.url = ckan.get_package_page_url(mdl.get_or_query_package_id(ckan))  # mark as source
         sample_mdl.package_attributes.state = CkanState.Active  # publish at the end of the process
+        if sample_groups is not None:
+            sample_mdl.package_attributes.groups = [ckan.get_group_info_or_request(group_name) for group_name in sample_groups]
         for resource_builder in sample_mdl.resource_builders.values():
             resource_builder.aliases = []
             if isinstance(resource_builder, BuilderDataStoreABC):